Move runner into scheduler/docker.
